+++ install-scripts/debian/20-debian-post-install-xorg.sh
@@ -540,7 +540,7 @@ Inherits=Breeze
 }
 
 #### DESKTOP ENV INSTALL
-if [ -d /sys/firmware/efi/efivars ] && sudo test ! -f "/boot/efi/startup.nsh"; then
+if [ -d /sys/firmware/efi/efivars ] && sudo test -d /boot/efi/EFI && sudo test ! -f /boot/efi/startup.nsh; then
   sudo mkdir -p /boot/efi/EFI/boot
   if [ -d "/boot/efi/EFI/refind" ]; then
     sudo cp -a /boot/efi/EFI/refind/refind_x64.efi /boot/efi/EFI/boot/bootx64.efi
