+++ install-scripts/fedora/10-rhel-post-install.sh
@@ -85,7 +85,7 @@ while true; do
   esac
 done
 
-if [ -d /sys/firmware/efi/efivars ] && sudo test ! -f "/boot/efi/startup.nsh"; then
+if [ -d /sys/firmware/efi/efivars ] && sudo test -d /boot/efi/EFI && sudo test ! -f /boot/efi/startup.nsh; then
   sudo mkdir -p /boot/efi/EFI/boot
   if [ -d "/boot/efi/EFI/refind" ]; then
     sudo cp -a /boot/efi/EFI/refind/refind_x64.efi /boot/efi/EFI/boot/bootx64.efi
